Refactor properly out of app into plugin
