+++ src/lib/components/App.svelte
@@ -17,16 +17,13 @@
 	import Details from '$lib/components/overlay/Details.svelte'
 	import TreeContainer from '$lib/components/overlay/left-pane/TreeContainer.svelte'
 	import Settings from '$lib/components/overlay/settings/Settings.svelte'
-	import { backendIP, websocketPort } from '$lib/defines'
 	import { provideWorld, traits, useQuery } from '$lib/ecs'
-	import { provideAnthropicKey, useAnthropicKey } from '$lib/hooks/useAnthropicKey.svelte'
 	import { type CameraPose, provideCameraControls } from '$lib/hooks/useControls.svelte'
 	import { provideEnvironment } from '$lib/hooks/useEnvironment.svelte'
 	import { providePartConfig } from '$lib/hooks/usePartConfig.svelte'
 	import { createPartIDContext } from '$lib/hooks/usePartID.svelte'
 	import { provideSettings } from '$lib/hooks/useSettings.svelte'
 	import { provideWeblabs } from '$lib/hooks/useWeblabs.svelte'
-	import { type InferCallback, LLMSceneBuilder } from '$lib/plugins'
 	import { domPortal } from '$lib/portal'
 
 	import FileDrop from './FileDrop/FileDrop.svelte'
@@ -79,13 +76,6 @@
 		 * Snippet to inject items into the details panel
 		 */
 		details?: Snippet<[{ entity: Entity }]>
-
-		/**
-		 * When provided, mounts the LLM Frame Builder plugin using this callback for inference.
-		 * In the standalone visualizer, the plugin mounts automatically with the LLM endpoint
-		 * in the draw-server. In the embedded app, pass in the org-scoped inference instead.
-		 */
-		onInfer?: InferCallback
 	}
 
 	let {
@@ -93,7 +83,6 @@
 		inputBindingsEnabled = true,
 		localConfigProps,
 		cameraPose,
-		onInfer,
 		settingsTabs,
 		children: appChildren,
 		dashboard,
@@ -111,27 +100,8 @@
 	provideCameraControls(() => cameraPose)
 	createPartIDContext(() => partID)
 
-	provideAnthropicKey()
 	provideWeblabs()
 	provideToast()
-
-	const anthropicKey = useAnthropicKey()
-
-	const standaloneInfer: InferCallback = async (prompt, components) => {
-		const res = await fetch(`http://${backendIP}:${websocketPort}/scene-builder`, {
-			method: 'POST',
-			headers: { 'Content-Type': 'application/json' },
-			body: JSON.stringify({
-				prompt,
-				components,
-				anthropicApiKey: anthropicKey.current || undefined,
-			}),
-		})
-		if (!res.ok) {
-			throw new Error(`${res.status}: ${await res.text()}`)
-		}
-		return res.json()
-	}
 
 	let root = $state.raw<HTMLElement>()
 
@@ -205,9 +175,6 @@
 				<Settings {settingsTabs} />
 				<Logs />
 				<AddFrames />
-				{#if !localConfigProps || onInfer}
-					<LLMSceneBuilder onInfer={onInfer ?? standaloneInfer} />
-				{/if}
 			</div>
 		</SceneProviders>
 	</Canvas>
