+++ src/lib/plugins/LLMSceneBuilder/SceneBuilder.svelte
@@ -12,6 +12,7 @@
 	let prompt = $state('')
 
 	const canSubmit = $derived(prompt.trim().length > 0 && sceneBuilder.uiState === 'idle')
+	$inspect(isOpen, 'isOpen')
 </script>
 
 <Portal id="dashboard">
