+++ src/lib/plugins/index.ts
@@ -19,6 +19,7 @@ export { default as Focus } from './Focus/Focus.svelte'
 
 // LLMSceneBuilder
 export { default as LLMSceneBuilder } from './LLMSceneBuilder/LLMSceneBuilder.svelte'
+export { default as AISettings } from './LLMSceneBuilder/AISettings.svelte'
 export type { InferCallback, ComponentFrameInfo } from './LLMSceneBuilder/useSceneBuilder.svelte'
 export type { FrameDelta } from './LLMSceneBuilder/frameDeltaAdapter'
